requirements.txt
py-cord
python-dotenv
amulet-nbt
yt-dlp
PyNaCl
async-timeout